LayoutTests/fast/events/event-leak-objects-expected.txt

PASS checkAllPropertiesFromCurrentWorld(documentResult) is true
PASS document.pageDefinedVar is undefined.
PASS checkAllPropertiesFromCurrentWorld(documentResult) is true
PASS document.pageDefinedVar is undefined.
PASS checkAllPropertiesFromCurrentWorld(documentResult) is true
PASS document.pageDefinedVar is undefined.
PASS checkAllPropertiesFromCurrentWorld(documentResult) is true
PASS document.pageDefinedVar is undefined.
PASS checkAllPropertiesFromCurrentWorld(documentResult) is true
PASS document.pageDefinedVar is undefined.
PASS checkAllPropertiesFromCurrentWorld(documentResult) is true
PASS document.pageDefinedVar is undefined.
PASS checkAllPropertiesFromCurrentWorld(documentResult) is true
PASS document.pageDefinedVar is undefined.
PASS checkAllPropertiesFromCurrentWorld(documentResult) is true
PASS document.pageDefinedVar is undefined.
PASS checkAllPropertiesFromCurrentWorld(documentResult) is true
PASS document.pageDefinedVar is undefined.
PASS checkAllPropertiesFromCurrentWorld(documentResult) is true
PASS document.pageDefinedVar is undefined.
PASS checkAllPropertiesFromCurrentWorld(documentResult) is true
PASS document.pageDefinedVar is undefined.
PASS checkAllPropertiesFromCurrentWorld(documentResult) is true
PASS document.pageDefinedVar is undefined.
PASS checkAllPropertiesFromCurrentWorld(documentResult) is true
PASS document.pageDefinedVar is undefined.
PASS checkAllPropertiesFromCurrentWorld(documentResult) is true
PASS document.pageDefinedVar is undefined.
PASS checkAllPropertiesFromCurrentWorld(documentResult) is true
PASS document.pageDefinedVar is undefined.
PASS checkAllPropertiesFromCurrentWorld(documentResult) is true
PASS document.pageDefinedVar is undefined.
PASS checkAllPropertiesFromCurrentWorld(documentResult) is true
PASS document.pageDefinedVar is undefined.
PASS checkAllPropertiesFromCurrentWorld(documentResult) is true
PASS document.pageDefinedVar is undefined.
PASS successfullyParsed is true

TEST COMPLETE
